+++ SleepHunter/Macro/PlayerMacroState.cs
@@ -25,11 +25,16 @@ namespace SleepHunter.Macro
 
         private int spellQueueIndex;
         private int flowerQueueIndex;
+
+        private SpellRotationMode spellQueueRotation;
         private bool isWaitingOnMana;
         private bool useLyliacVineyard;
         private bool flowerAlternateCharacters;
+        private bool skipSpellsOnCooldown;
+
         private DateTime spellCastTimestamp;
         private TimeSpan spellCastDuration;
+
         private SpellQueueItem lastUsedSpellItem;
         private SpellQueueItem fasSpioradQueueItem;
         private SpellQueueItem lyliacPlantQueueItem;
@@ -61,22 +66,34 @@ namespace SleepHunter.Macro
 
         public int FlowerQueueCount => flowerQueue.Count;
 
+        public SpellRotationMode SpellQueueRotation
+        {
+            get => spellQueueRotation;
+            set => SetProperty(ref spellQueueRotation, value);
+        }
+
         public bool IsWaitingOnMana
         {
             get => isWaitingOnMana;
-            set => SetProperty(ref isWaitingOnMana, value, nameof(IsWaitingOnMana));
+            set => SetProperty(ref isWaitingOnMana, value);
         }
 
         public bool UseLyliacVineyard
         {
             get => useLyliacVineyard;
-            set => SetProperty(ref useLyliacVineyard, value, nameof(UseLyliacVineyard));
+            set => SetProperty(ref useLyliacVineyard, value);
         }
 
         public bool FlowerAlternateCharacters
         {
             get => flowerAlternateCharacters;
-            set => SetProperty(ref flowerAlternateCharacters, value, nameof(FlowerAlternateCharacters));
+            set => SetProperty(ref flowerAlternateCharacters, value);
+        }
+
+        public bool SkipSpellsOnCooldown
+        {
+            get => skipSpellsOnCooldown;
+            set => SetProperty(ref skipSpellsOnCooldown, value);
         }
 
         public DateTime SpellCastTimestamp
@@ -437,7 +454,7 @@ namespace SleepHunter.Macro
                 {
                     if (flowerQueue.Count > 0)
                         SetPlayerStatus(PlayerMacroStatus.ReadyToFlower);
-                    else if (client.Skillbook.ActiveSkills.Count() > 0)
+                    else if (client.Skillbook.ActiveSkills.Any())
                         SetPlayerStatus(PlayerMacroStatus.Waiting);
                     else
                         SetPlayerStatus(PlayerMacroStatus.Idle);
@@ -827,34 +844,22 @@ namespace SleepHunter.Macro
 
         private SpellQueueItem GetNextSpell()
         {
-            client.Update(PlayerFieldFlags.Spellbook);
-
-            var shouldRotate = UserSettingsManager.Instance.Settings.SpellRotationMode != SpellRotationMode.None;
-            var isRoundRobin = UserSettingsManager.Instance.Settings.SpellRotationMode == SpellRotationMode.RoundRobin;
-
-            if (spellQueueIndex >= spellQueue.Count)
-                spellQueueIndex = 0;
-
             if (spellQueue.Count < 1)
                 return null;
 
-            var currentSpell = spellQueue.ElementAt(spellQueueIndex);
-            var currentId = currentSpell.Id;
+            client.Update(PlayerFieldFlags.Spellbook);
 
-            while (currentSpell.IsDone && shouldRotate)
+            var skipOnCooldown = UserSettingsManager.Instance.Settings.SkipSpellsOnCooldown;
+
+            var currentSpell = SpellQueueRotation switch
             {
-                if (++spellQueueIndex >= spellQueue.Count)
-                    spellQueueIndex = 0;
+                SpellRotationMode.None => GetNextSpell_NoRotation(skipOnCooldown),
+                SpellRotationMode.Singular => GetNextSpell_SingularOrder(skipOnCooldown),
+                SpellRotationMode.RoundRobin => GetNextSpell_RoundRobin(skipOnCooldown),
+                _ => null,
+            };
 
-                currentSpell = spellQueue.ElementAt(spellQueueIndex);
-
-                if (currentSpell.Id == currentId)
-                {
-                    IsWaitingOnMana = false;
-                    return null;
-                }
-            }
-
+            // Determine if we need to fas spiorad instead
             if (currentSpell != null)
             {
                 var currentSpellData = SpellMetadataManager.Instance.GetSpell(currentSpell.Name);
@@ -863,10 +868,52 @@ namespace SleepHunter.Macro
                     return GetFasSpiorad();
             }
 
-            if (isRoundRobin && shouldRotate)
-                spellQueueIndex++;
+            if (currentSpell.IsOnCooldown)
+                return null;
 
             return !currentSpell.IsDone ? currentSpell : null;
+        }
+
+        private SpellQueueItem GetNextSpell_NoRotation(bool skipOnCooldown = true)
+        {
+            if (skipOnCooldown)
+                return spellQueue.FirstOrDefault(spell => !spell.IsOnCooldown);
+            else
+                return spellQueue.FirstOrDefault();
+        }
+
+        private SpellQueueItem GetNextSpell_SingularOrder(bool skipOnCooldown = true)
+        {
+            if (skipOnCooldown)
+                return spellQueue.FirstOrDefault(spell => !spell.IsOnCooldown && !spell.IsDone);
+            else
+                return spellQueue.FirstOrDefault(spell => !spell.IsDone);
+        }
+
+        private SpellQueueItem GetNextSpell_RoundRobin(bool skipOnCooldown = true)
+        {
+            // All spells are done, nothing to cast
+            if (spellQueue.All(spell => spell.IsDone))
+                return null;
+
+            // All spells are on cooldown, and skipping so nothing to do
+            if (spellQueue.All(spell => spell.IsOnCooldown) && skipOnCooldown)
+                return null;
+
+            var currentSpell = spellQueue.ElementAt(spellQueueIndex);
+
+            while (currentSpell.IsDone || (skipOnCooldown && currentSpell.IsOnCooldown))
+                currentSpell = AdvanceToNextSpell();
+
+            // Round robin rotation for next time
+            AdvanceToNextSpell();
+            return currentSpell;
+        }
+
+        private SpellQueueItem AdvanceToNextSpell()
+        {
+            spellQueueIndex = (spellQueueIndex + 1) % spellQueue.Count;
+            return spellQueue.Count > 0 ? spellQueue[spellQueueIndex] : null;
         }
 
         private FlowerQueueItem GetNextFlowerTarget()
@@ -1053,7 +1100,7 @@ namespace SleepHunter.Macro
             if (item == null)
                 throw new ArgumentNullException(nameof(item));
 
-            client.Update(PlayerFieldFlags.Inventory | PlayerFieldFlags.Equipment | PlayerFieldFlags.Stats);
+            client.Update( PlayerFieldFlags.Inventory | PlayerFieldFlags.Equipment | PlayerFieldFlags.Stats | PlayerFieldFlags.Spellbook);
 
             var equippedStaff = client.Equipment.GetSlot(EquipmentSlot.Weapon);
             var availableList = new List<string>(client.Inventory.ItemNames);