+++ SleepHunter/Macro/SpellRotationMode.cs
@@ -3,6 +3,7 @@ namespace SleepHunter.Macro
 {
     public enum SpellRotationMode
     {
+        Default,
         None,
         Singular,
         RoundRobin