+++ SleepHunter/Views/MainWindow.xaml.cs
@@ -555,12 +555,19 @@ namespace SleepHunter.Views
                 macro.Client.PlayerUpdated += HandleClientUpdateTick;
             }
 
+            var didLoadFromState = false;
+
             try
             {
                 if (shouldRecallMacroState && macro != null)
                 {
                     logger.LogInfo($"Attempting to load previous macro state for character: {player.Name}");
                     LoadMacroState(player);
+
+                    if (macro.SpellQueueRotation == SpellRotationMode.Default)
+                        macro.SpellQueueRotation = UserSettingsManager.Instance.Settings.SpellRotationMode;
+
+                    didLoadFromState = true;
                 }
             }
             catch (Exception ex)
@@ -576,10 +583,15 @@ namespace SleepHunter.Views
                    MessageBoxButton.OK, 460, 260);
 
                 }, DispatcherPriority.Normal, null);
+
             }
             finally
             {
                 UpdateWindowTitle();
+
+                // Set default spell queue rotation mode
+                if (!didLoadFromState)
+                    macro.SpellQueueRotation = UserSettingsManager.Instance.Settings.SpellRotationMode;
             }
         }
 
@@ -671,6 +683,7 @@ namespace SleepHunter.Views
 
                 queuedSpell.MaximumLevel = spell.MaximumLevel;
                 queuedSpell.CurrentLevel = spell.CurrentLevel;
+                queuedSpell.IsOnCooldown = spell.IsOnCooldown;
             }
         }
 
@@ -1503,7 +1516,7 @@ namespace SleepHunter.Views
             if (selectedMacro == null || selectedMacro.Client == null || !selectedMacro.Client.IsLoggedIn)
                 return;
 
-            selectedMacro.Client.Update(PlayerFieldFlags.All);
+            selectedMacro.Client.Update();
             selectedMacro.Start();
             UpdateToolbarState();
 
@@ -1804,6 +1817,8 @@ namespace SleepHunter.Views
 
             if (selectedMacro != null)
             {
+                spellQueueRotationComboBox.SelectedValue = selectedMacro.SpellQueueRotation;
+
                 spellQueueListBox.ItemsSource = selectedMacro.QueuedSpells;
                 RefreshSpellQueue();
 
@@ -2159,33 +2174,33 @@ namespace SleepHunter.Views
 
             logger.LogInfo($"User setting property changed: {e.PropertyName}");
 
-            if (string.Equals("SelectedTheme", e.PropertyName, StringComparison.OrdinalIgnoreCase))
+            if (string.Equals(nameof(settings.SelectedTheme), e.PropertyName, StringComparison.OrdinalIgnoreCase))
                 ApplyTheme();
 
-            if (string.Equals("ClientSortOrder", e.PropertyName, StringComparison.OrdinalIgnoreCase))
+            if (string.Equals(nameof(settings.ClientSortOrder), e.PropertyName, StringComparison.OrdinalIgnoreCase))
             {
                 PlayerManager.Instance.SortOrder = settings.ClientSortOrder;
                 UpdateClientList();
             }
 
-            if (string.Equals("SkillGridWidth", e.PropertyName, StringComparison.OrdinalIgnoreCase))
+            if (string.Equals(nameof(settings.SkillGridWidth), e.PropertyName, StringComparison.OrdinalIgnoreCase))
                 SetSkillGridWidth(settings.SkillGridWidth);
 
-            if (string.Equals("WorldSkillGridWidth", e.PropertyName, StringComparison.OrdinalIgnoreCase))
+            if (string.Equals(nameof(settings.WorldSkillGridWidth), e.PropertyName, StringComparison.OrdinalIgnoreCase))
                 SetWorldSkillGridWidth(settings.WorldSkillGridWidth);
 
-            if (string.Equals("SpellGridWidth", e.PropertyName, StringComparison.OrdinalIgnoreCase))
+            if (string.Equals(nameof(settings.SpellGridWidth), e.PropertyName, StringComparison.OrdinalIgnoreCase))
                 SetSpellGridWidth(settings.SpellGridWidth);
 
-            if (string.Equals("WorldSpellGridWidth", e.PropertyName, StringComparison.OrdinalIgnoreCase))
+            if (string.Equals(nameof(settings.WorldSpellGridWidth), e.PropertyName, StringComparison.OrdinalIgnoreCase))
                 SetWorldSpellGridWidth(settings.WorldSpellGridWidth);
 
-            if (string.Equals("SkillIconSize", e.PropertyName, StringComparison.OrdinalIgnoreCase))
+            if (string.Equals(nameof(settings.SkillIconSize), e.PropertyName, StringComparison.OrdinalIgnoreCase))
                 UpdateSkillSpellGridWidths();
 
             // Debug settings
 
-            if (string.Equals("ShowAllProcesses", e.PropertyName, StringComparison.OrdinalIgnoreCase))
+            if (string.Equals(nameof(settings.ShowAllProcesses), e.PropertyName, StringComparison.OrdinalIgnoreCase))
             {
                 PlayerManager.Instance.ShowAllClients = settings.ShowAllProcesses;
                 UpdateClientList();
@@ -2197,8 +2212,26 @@ namespace SleepHunter.Views
             if (sender is not PlayerMacroState macro)
                 return;
 
-            if (string.Equals("Status", e.PropertyName, StringComparison.OrdinalIgnoreCase))
+            if (string.Equals(nameof(macro.Status), e.PropertyName, StringComparison.OrdinalIgnoreCase))
                 UpdateUIForMacroStatus(macro.Status);
+
+            // Update Spell Queue Rotation
+            if (string.Equals(nameof(macro.SpellQueueRotation), e.PropertyName, StringComparison.OrdinalIgnoreCase))
+                spellQueueRotationComboBox.SelectedValue = macro.SpellQueueRotation;
+        }
+
+        private void spellQueueRotationComboBox_SelectionChanged(object sender, SelectionChangedEventArgs e)
+        {
+            if (selectedMacro == null || e.AddedItems.Count < 1)
+                return;
+
+            if (e.AddedItems[0] is not UserSetting selection)
+                return;
+
+            if (!Enum.TryParse<SpellRotationMode>(selection.Value as string, out var newRotationMode))
+                return;
+
+            selectedMacro.SpellQueueRotation = newRotationMode;
         }
 
         private void flowerVineyardCheckBox_Checked(object sender, RoutedEventArgs e)